Rebase server & announcements branches. Aggregate types into barreled types directory. Setup API communication from frontend. Add axios & react-query for fetch support. Resolve CORS issue for localhost.
