+++ frontend/package.json
@@ -13,6 +13,8 @@
   "dependencies": {
     "@mantine/core": "^7.15.2",
     "@mantine/hooks": "^7.15.2",
+    "@tanstack/react-query": "^5.64.1",
+    "axios": "^1.7.9",
     "motion": "^11.18.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
