+++ frontend/src/App.tsx
@@ -5,13 +5,18 @@ import '@mantine/core/styles.css'
 import { KissenaTheme, KissenaCSSResolver } from '@/kissena/theme'
 import '@kissena/theme.module.css'
 
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import Router from '@/Router'
+
+const queryClient = new QueryClient();
 
 function App() {
   return (
-    <MantineProvider theme={KissenaTheme} cssVariablesResolver={KissenaCSSResolver}>
-      <Router />
-    </MantineProvider>
+    <QueryClientProvider client={queryClient}>
+      <MantineProvider theme={KissenaTheme} cssVariablesResolver={KissenaCSSResolver}>
+        <Router />
+      </MantineProvider>
+    </QueryClientProvider>
   )
 }
