+++ frontend/src/api/config.ts
@@ -1,0 +1,1 @@
+export const HOSTNAME = "http://localhost:3000";
