+++ frontend/src/kissena/components/Announcement/Announcement.tsx
@@ -1,13 +1,7 @@
 import styles from'@kissena/components/Announcement/Announcement.module.css'
-import { Box, Button, Space, Stack, Text } from '@mantine/core'
+import { Box, Button, Space, Stack, Text } from '@mantine/core' 
 import { useMemo } from 'react'
-
-export interface AnnouncementData {
-    title:string,
-    datePosted:Date
-    description?:string
-    redirectUrl?:string
-};
+import { AnnouncementData } from '@/types';
 
 export function Announcement({title, datePosted, description, redirectUrl}:AnnouncementData) { 
     // Configure width of learn more button (and wrapper anchor tag)
@@ -21,9 +15,7 @@ export function Announcement({title, datePosted, description, redirectUrl}:Annou
     return (
         <Stack gap={0} pt={'lg'} pb={'xl'} className={styles.announcementContainer}>
             <Text size='md' c="neonGreen.6" fw={700} lh={'1.2rem'} lineClamp={2}>{title}</Text>
-            <Text size='md' c="neonGreen.9">{stringifiedDate}</Text>
-
-            
+            <Text size='md' c="neonGreen.9">{stringifiedDate}</Text>            
             {
                 !description
                 ? <></>
