+++ frontend/src/types/AnnouncementData.ts
@@ -1,0 +1,6 @@
+export interface AnnouncementData {
+    title:string,
+    datePosted:Date
+    description?:string
+    redirectUrl?:string
+};
