+++ frontend/src/types/index.ts
@@ -1,0 +1,3 @@
+import { AnnouncementData } from "./AnnouncementData";
+
+export type { AnnouncementData };
