+++ frontend/vite.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
     alias: {
       '@': path.resolve(__dirname, 'src/'),
       '@kissena': path.resolve(__dirname, 'src/kissena'),
+      '@api': path.resolve(__dirname, 'src/api'),
       '@admin': path.resolve(__dirname, 'src/admin'),
     },
   },
