+++ server/api/functions/announcements/getAnnouncementCount.ts
@@ -2,5 +2,5 @@ import { Request, Response } from "express";
 
 // Used for GET: /announcements/count
 export function getAnnouncementCount(req: Request, res: Response) {
-  res.send("0");
+  res.send("10");
 }
