+++ server/package.json
@@ -11,9 +11,9 @@
     "lint:fix": "eslint . --fix "
   },
   "_moduleAliases": {
-    "@api" : "api",
-    "@routes" : "api/routes",
-    "@functions" : "api/functions"
+    "@api": "api",
+    "@routes": "api/routes",
+    "@functions": "api/functions"
   },
   "keywords": [],
   "author": "Patina Network",
@@ -21,6 +21,7 @@
   "type": "commonjs",
   "description": "The backend server for running Kissena Synergy's website.",
   "dependencies": {
+    "cors": "^2.8.5",
     "express": "^4.21.2"
   },
   "devDependencies": {
